Alphabetize contacts list in GUI
