+++ client/client.go
@@ -70,7 +70,7 @@ import (
 	"io/ioutil"
 	"net"
 	"os"
-	"os/exec"
+	"sort"
 	"strconv"
 	"sync"
 	"time"
@@ -214,9 +214,6 @@ type client struct {
 	// disableV2Ratchet causes the client to advertise and process V1
 	// axolotl ratchet support.
 	disableV2Ratchet bool
-
-	// command to run upon receiving messages
-	receiveHookCommand string
 }
 
 // UI abstracts behaviour that is specific to a given interface (GUI or CLI).
@@ -484,6 +481,30 @@ type Event struct {
 	msg string
 }
 
+// clientList is a sortable listing of contacts
+type contactList []*Contact
+
+func (cl contactList) Len() int {
+	return len(cl)
+}
+
+func (cl contactList) Less(i, j int) bool {
+	return cl[i].name < cl[j].name
+}
+
+func (cl contactList) Swap(i, j int) {
+	cl[i], cl[j] = cl[j], cl[i]
+}
+
+func (c *client) ContactsSorted() ([]*Contact) {
+	contacts := contactList(make([]*Contact, 0, len(c.contacts)))
+	for i := range c.contacts {
+		contacts = append(contacts, c.contacts[i])
+	}
+	sort.Sort(contacts)
+	return contacts
+}
+
 // previousTagLifetime contains the amount of time that we'll store a previous
 // tag (or previous group private key) for.
 const previousTagLifetime = 14 * 24 * time.Hour
@@ -731,8 +752,6 @@ func (c *client) loadUI() error {
 			return err
 		}
 	}
-
-	c.receiveHookCommand = os.Getenv("POND_HOOK_RECEIVE")
 
 	c.ui.loadingUI()
 
@@ -1418,15 +1437,4 @@ func (c *client) importTombFile(stateFile *disk.StateFile, keyHex, path string) 
 	<-writerDone
 
 	return nil
-}
-
-func (c *client) receiveHook() {
-	if c.receiveHookCommand != "" {
-		cmd := exec.Command(c.receiveHookCommand)
-		go func() {
-			if err := cmd.Run(); err != nil {
-				c.log.Errorf("Failed to run receive hook command: %s", err.Error())
-			}
-		}()
-	}
 }
